event pageの本体をコンポーネントに切り出した
